Fixed the Background after merge
