+++ src/App.css
@@ -13,7 +13,6 @@ html {
   width: 100%;
   height: 100%;
   z-index: -1;
-  /* Ensure it stays behind other content */
 }
